Added basic script to get transition matrices
The script is extremely simple. It looks at a character in a text file (starting from the second character) and compares it to the previous character. It records the number of times the character is preceded by a character in a matrix T. It then exports the matrix as a .csv file.